requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
docopt==0.6.2
filelock==3.12.0
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
fonttools==4.39.4
greenlet==2.0.2
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
matplotlib==3.7.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
prompt-toolkit==3.0.39
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
redis==4.6.0
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.17
systemd-python==234
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
vine==5.0.0
virtualenv==20.23.0
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.3.4
yarg==0.1.9
zipp==1.0.0
